deps: remove bitcoin-spv
