+++ bitcoins/Cargo.toml
@@ -8,7 +8,6 @@ repository = "https://github.com/summa-tx/bitcoins-rs"
 license = "MIT OR Apache-2.0"
 
 [dependencies]
-bitcoin-spv = "5.0.0"
 hex = "0.4.2"
 bech32 = "0.7.2"
 base58check = "0.1.0"
