+++ bitcoins/src/prelude.rs
@@ -7,7 +7,5 @@ pub use crate::{
 
 pub use coins_core::prelude::*;
 
-pub use bitcoin_spv::types::RawHeader;
-
 #[cfg(any(feature = "mainnet", feature = "testnet", feature = "signet"))]
 pub use crate::defaults::*;
